Extract StatusOptions to a separate file
